// src/Types/types.ts
export interface TLandDetails {
    id: string;
    name: string;
    location: string;
    description: string;
    polygon: string[];
    initialCood: string[];
    defaultZooming: number;
    images: string[];
    videos: string[];
    plotNumber: string;
    size: number;
    purpose: string;
    security: string;
    documentation: string;
    environment: string;
    allodialOwnership: string;
    ground: string;
    price: number;
    etaToCBD: string;

}